Document와 Root폴더 하위에 Container과 Component 분리.
