+++ src/App.js
@@ -12,7 +12,6 @@ export default function App({ $app }) {
 
   this.setState = (nextState) => {
     this.state = nextState;
-    console.log("app.this.state", this.state);
     rootContainer.setState(this.state);
   };
 
@@ -23,40 +22,10 @@ export default function App({ $app }) {
     this.setState(roots);
   };
 
-  fetchRoot();
-
   const fetchDoc = async (id) => {
     const doc = await request(`/documents/${id}`);
     return doc;
   };
-
-  const rootContainer = new RootContainer({
-    $target,
-    initialState,
-    onClick: async (documentId) => {
-      const doc = await fetchDoc(documentId);
-      documentContainer.setState(doc);
-      documentContainer.render();
-    },
-    addNewDoc: async () => {
-      console.log("addneDoc시작");
-      const res = await fetchNewPost();
-      //{ res형식
-      // "id": 6,
-      // "title": "문서 제목",
-      // "createdAt": "생성된 날짜",
-      // "updatedAt": "수정된 날짜"
-      //}
-      //이 res를 root에 추가해줘
-      await fetchRoot(); //루트는 새로 렌더링해주고
-      //이제 fetchDoc해줘야지.
-      console.log(res); //res에 대해 출력됐고
-      const doc = await fetchDoc(res.id);
-      documentContainer.setState(doc);
-      documentContainer.render();
-    }, //Root에 새 문서추가해주면서 새 문서 작성 페이지로 이동해야대
-    //됐는데 내용이 null로 나오네
-  });
 
   const fetchNewPost = async () => {
     const res = await request(`/documents`, {
@@ -68,17 +37,25 @@ export default function App({ $app }) {
     });
     return res;
   };
+  fetchRoot();
 
-  const EditDoc = async (document) => {
-    const res = await request(`/documents/${document.id}`, {
-      method: "PUT",
-      body: JSON.stringify({
-        title: `${document.title}`,
-        content: `${document.content}`,
-      }),
-    });
-  };
-
+  const rootContainer = new RootContainer({
+    $target,
+    initialState,
+    onRenderDoc: async (documentId) => {
+      const doc = await fetchDoc(documentId); //doc정보 보내줘
+      documentContainer.setState(doc);
+      //documentContainer.render();
+    },
+    onAddDoc: async () => {
+      const res = await fetchNewPost();
+      await fetchRoot(); //루트는 새로 렌더링해주고
+      const doc = await fetchDoc(res.id);
+      documentContainer.setState(doc);
+      //documentContainer.render();
+    }, //Root에 새 문서추가해주면서 새 문서 작성 페이지로 이동해야대
+    //됐는데 내용이 null로 나오네
+  });
   const documentContainer = new DocumentContainer({
     $target,
     initialState: {
@@ -88,10 +65,6 @@ export default function App({ $app }) {
       documents: [],
       createAt: null,
       updatedAt: null,
-    },
-    onEdit: (document) => {
-      console.log("온에딧", document);
-      EditDoc(document);
     },
   });
 }
