+++ src/Document/DocumentContainer.js
@@ -1,34 +1,34 @@
-export default function DocumentContainer({ $target, initialState, onEdit }) {
+import Editor from "./Editor.js";
+import { request } from "../../library/api.js";
+
+export default function DocumentContainer({ $target, initialState }) {
   const $page = document.createElement("div");
   $page.className = "DocumentContainer";
   $target.appendChild($page);
   $page.innerHTML = "DocumentContainer";
-
   this.state = initialState;
+
+  const EditDoc = async (document) => {
+    const res = await request(`/documents/${document.id}`, {
+      method: "PUT",
+      body: JSON.stringify({
+        title: `${document.title}`,
+        content: `${document.content}`,
+      }),
+    });
+  };
+
+  const $editor = new Editor({
+    $page,
+    initialState,
+    onEdit: (nextState) => {
+      EditDoc(nextState);
+    },
+  });
 
   this.setState = (nextState) => {
     this.state = nextState;
+    $editor.setState(nextState);
+    $editor.render();
   };
-
-  let isInit = false;
-
-  this.render = () => {
-    if (this.state.title !== null) {
-      $page.innerHTML = `
-    <input type="text" name="title" style="width:93%;  height:5%; margin:20px; " value=${this.state.title}>
-    <textarea name="content" style="width:93%; height:80%; margin:20px;">${this.state.content}</textarea>
-`;
-      isInit = true;
-    }
-  };
-
-  $page.addEventListener("keyup", (e) => {
-    const name = e.target.name;
-    const nextState = {
-      ...this.state,
-      [name]: e.target.value,
-    };
-    this.setState(nextState);
-    onEdit(nextState);
-  });
 }
